New stoneholmespassagetime and trywarning/catchwarning functions, Completely reworked stoneholmesargs, NSF acknowledgment and disclaimer
New stoneholmespassagetime function added to suite to calculate passage times given a Ston-Holmes parameter set. The function is fully vectorized. Try/catch of warnings supported via pair of functions: trywarning/catchwarning. This code is quite flexible and supports multiple warning, but is still a bit of a hack. May a class object would be better? The stoneholmesargs function was completely refactored and made much more efficient and hopeful easier to understand. A first draft of of an ismatrix substitute, shc_ismatrix is included. Finally, the NSF acknowledgment and disclaimer was added to the README and Contents.m files.

Signed-off-by: Andrew D. Horchler <adh9@case.edu>