Changed order of operations in integration function to better handle small values.
Integrating networks with no noise and no input (Mu) showed and plotting output activity vs. time on a semiology plot showed that the slopes changed as the trajectory converged on the SHC manifold. This is likely do to numeric cancellation of small values. Changing the order of operations in the integration function seems to have made the slopes the same - or at least much much closer to being identical.

Signed-off-by: Andrew D. Horchler <adh9@case.edu>