Much faster inter-passage transition time calculation and parameter finding
Adjusted integration method in shc_lv_transitiontime() to speed it up by an order of magnitude or so. Much faster. A simple approximation based on the transition time for a marginally stable system is is used to estimate step-size. Runge-Kutta now used for time-step counting with a simple step-size reduction scheme to converge on the transition time, Tt.

The new shc_lv_transitiontime() makes shc_lv_params much faster again. Additionally, a correction on the initial guess for Alpha was added based on shc_lv_transitiontime().

Signed-off-by: Andrew D. Horchler <adh9@case.edu>