Improved mean period constant-find function
Improved shc_lv_meanperiod(). The mean period constant is now found by fitting the mean period data with stoneholmesfit(). This potentially allows for smaller values of N and improve the stability of the output due to noise. The fitted parameters are irrelevant; they are just needed for use with stoneholmesfirstpassagetime(). Future improvements are possible.

Fixed off-by-one bug that caused function to ignore value of N. And set default N to 300 instead of 30.

Removed instanced of parlor for now due to Matlab bug that breaks parallel processing. Need solution that will be robust and not depend on user having a patched system.

Also a minor tweak to stoneholmesfit() to disable printing of output.

Signed-off-by: Andrew D. Horchler <adh9@case.edu>