Simplified connection matrix (rho) parametrization!
Removed Nu parameter from Gamma-like (non-diagonal, non i-1) elements of Rho. Now all unstable eigenvalues of node i are -Alpha_i (repeated eigenvalues). This is the weakest stability case for the "higher order" eigenvalues, but it simplifies the calculation of the saddle values and thus the stability when Nu < 1.

(It's conceivable that this could change back or that an option could be added to adjust the higher order terms.)

Signed-off-by: Andrew D. Horchler <adh9@case.edu>