Mu and Epsilon functions can now be a function of state, fixed bugs
Mu and Epsilon (noise) which can optionally be function handles can now be functions of the state variables. Note however, that noise is still treated as (integrated as such) additive, so this feature should not be used to implement multiplicative noise. In other words, Epsilon should not be a continuous function of the state, rather this state can be used to switch between modes.

Replaced two instances of sde_ismatrix() with shc_ismatrix();

Signed-off-by: Andrew D. Horchler <adh9@case.edu>